move RecoveryKeyFactory and WordEncoder to common.recovery
